pyproject.toml
[project]
name = "secretary"
version = "0.1.0"
description = "社交媒体内容分析助手"
authors = [
    {name = "Bpbby", email = "zkd8907@live.com"},
]
dependencies = [
    "python-dotenv",
    "redis",
    "requests",
    "beautifulsoup4",
    "tweety-ns",
    "truthbrush",
    "langchain-openai",
    "pytz",
    "googletrans>=4.0.2",
]
requires-python = ">=3.11"

[tool.uv.sources]
truthbrush = { git = "https://github.com/stanfordio/truthbrush.git" }
